+++ frontend/src/components/schedule/index.ts
@@ -4,10 +4,10 @@
  * Barrel export for schedule editing components.
  */
 
-export { ScheduleEditor } from "./ScheduleEditor";
-export { ScheduleToolbar } from "./ScheduleToolbar";
-export { ScheduleTable } from "./ScheduleTable";
-export { ScheduleTableRow } from "./ScheduleTableRow";
-export { MiniTimelineBar } from "./MiniTimelineBar";
 export { EdgeScheduleTable } from "./EdgeScheduleTable";
 export { EdgeScheduleTableRow } from "./EdgeScheduleTableRow";
+export { MiniTimelineBar } from "./MiniTimelineBar";
+export { ScheduleEditor } from "./ScheduleEditor";
+export { ScheduleTable } from "./ScheduleTable";
+export { ScheduleTableRow } from "./ScheduleTableRow";
+export { ScheduleToolbar } from "./ScheduleToolbar";